Moving some functions from index to state-state
Doesn't really change the complexity much, but tossing the functions on as partially-applied "methods" of the state holder does clean up index.js a bit.

Alternately, they could have all gone to a shared utility file or something, but I'm not sure there would have been any advantage there.
